Moved around the GeoIP2Adapter
