+++ .travis.yml
@@ -3,7 +3,7 @@ sudo: false
 addons:
   apt:
     packages:
-    - libgeoip-dev
+      - libgeoip-dev
 
 language: php
 
@@ -26,7 +26,6 @@ matrix:
 
 before_script:
     - if [ "$TRAVIS_PHP_VERSION" != "hhvm" && "$TRAVIS_PHP_VERSION" != "7.0" ]; then pecl install geoip; fi
-    - composer self-update
     - if [ "$deps" = "low" ]; then composer update --prefer-dist --prefer-lowest; fi
     - if [ "$deps" = "" ]; then composer install --prefer-dist --no-interaction; fi
     - if [ "$TRAVIS_PHP_VERSION" == "7.0" ]; then composer require "geoip/geoip"; fi
